Use DateComponentOrdering as an options parameter Moved part of the code from TemporalInterpreter.java to TemporalParser.java
