+++ gbif/pipelines/ingest-gbif-java/src/main/java/org/gbif/pipelines/ingest/java/pipelines/InterpretedToEsIndexPipeline.java
@@ -142,7 +142,7 @@ public class InterpretedToEsIndexPipeline {
     BasicTransform basicTransform = BasicTransform.builder().create();
     MetadataTransform metadataTransform = MetadataTransform.builder().create();
     VerbatimTransform verbatimTransform = VerbatimTransform.create();
-    TemporalTransform temporalTransform = TemporalTransform.create();
+    TemporalTransform temporalTransform = TemporalTransform.builder().create();
     TaxonomyTransform taxonomyTransform = TaxonomyTransform.builder().create();
     GrscicollTransform grscicollTransform = GrscicollTransform.builder().create();
     LocationTransform locationTransform = LocationTransform.builder().create();
