+++ sdks/beam-common/src/main/java/org/gbif/pipelines/common/beam/options/InterpretationPipelineOptions.java
@@ -9,6 +9,7 @@ import org.apache.beam.sdk.options.Default;
 import org.apache.beam.sdk.options.DefaultValueFactory;
 import org.apache.beam.sdk.options.Description;
 import org.apache.beam.sdk.options.PipelineOptions;
+import org.gbif.common.parsers.date.DateComponentOrdering;
 
 /**
  * Pipeline options (configuration) for GBIF based data interpretation pipelines. Optionally can use
@@ -90,11 +91,10 @@ public interface InterpretationPipelineOptions
    * invalid, only accepts standard ISO date format. Parsing 2/3/2008 will fail . <code>DMY</code>
    * will parse 2/3/2008 as 2 Mar 2008 <code>MDY</code> will parse 2/3/2008 as 3 Feb 2008
    */
-  //  @Description("Recommended date formats for ambiguous dates, e.g. 2/3/2008")
-  //  @Default.String("ISO")
-  //  String getDefaultDateFormat();
-  //
-  //  void setDefaultDateForamt(String  defaultDateFormat);
+  @Description("Recommended date formats for ambiguous dates, e.g. 2/3/2008")
+  DateComponentOrdering getDefaultDateFormat();
+
+  void setDefaultDateFormat(DateComponentOrdering defaultDateFormat);
 
   /** A {@link DefaultValueFactory} which locates a default directory. */
   class TempDirectoryFactory implements DefaultValueFactory<String> {
